+++ src/check.rs
@@ -304,19 +304,8 @@ impl<'a> VirtualEnv<'a> {
             Instr::Prim(prim, _) => match prim {
                 Reduce => {
                     let sig = self.pop_func()?.signature();
-                    let outputs = match (sig.args, sig.outputs) {
-                        (0, _) => return Err(format!("{prim}'s function has no args").into()),
-                        (1, 0) => 0,
-                        (1, _) => {
-                            return Err(SigCheckError::from(format!(
-                                "{prim}'s function's signature is {sig}"
-                            ))
-                            .ambiguous())
-                        }
-                        (_, 1) => 1,
-                        _ => return Err(format!("{prim}'s function's signature is {sig}").into()),
-                    };
-                    self.handle_args_outputs(1, outputs)?;
+                    let args = sig.args.saturating_sub(sig.outputs);
+                    self.handle_args_outputs(args, sig.outputs)?;
                 }
                 Scan => {
                     let _sig = self.pop_func()?.signature();
@@ -541,21 +530,10 @@ impl<'a> VirtualEnv<'a> {
                     self.handle_args_outputs(args, outputs)?;
                 }
             },
-            Instr::ImplPrim(prim @ ImplPrimitive::ReduceContent, _) => {
+            Instr::ImplPrim(ImplPrimitive::ReduceContent | ImplPrimitive::ReduceDepth(_), _) => {
                 let sig = self.pop_func()?.signature();
-                let outputs = match (sig.args, sig.outputs) {
-                    (0, _) => return Err(format!("{prim}'s function has no args").into()),
-                    (1, 0) => 0,
-                    (1, _) => {
-                        return Err(SigCheckError::from(format!(
-                            "{prim}'s function's signature is {sig}"
-                        ))
-                        .ambiguous())
-                    }
-                    (_, 1) => 1,
-                    _ => return Err(format!("{prim}'s function's signature is {sig}").into()),
-                };
-                self.handle_args_outputs(1, outputs)?;
+                let args = sig.args.saturating_sub(sig.outputs);
+                self.handle_args_outputs(args, sig.outputs)?;
             }
             Instr::ImplPrim(prim, _) => {
                 let args = prim.args();